ml-services/feedback-generator/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# NLP and Text Generation
transformers>=4.37.2
torch>=2.2.0
openai==1.12.0

# Utils
python-dotenv==1.0.1
google-generativeai==0.4.1
